requirements.txt
# Core Python Dependencies
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
gunicorn>=20.1.0,<21.0.0
pydantic>=1.10.0,<1.11.0

# Agentic Framework
autogen>=0.2.0
python-dotenv>=1.0.0

# Async and Multiprocessing
asyncio>=3.4.3,<3.5.0

# Numerical and Scientific Computing
numpy>=1.24.3,<1.25.0
pandas>=2.0.1,<2.1.0
scipy>=1.10.1,<1.11.0
scikit-learn>=1.2.0,<1.3.0

# Trading and Financial Libraries
ccxt>=2.6.16,<2.7.0

# Quantum Computing & AI
pennylane>=0.30.0
qiskit>=0.40.0
tensorflow>=2.12.0
torch>=2.0.0
transformers>=4.30.0
prophet>=1.1.2

# Quantum Computing Simulation
qutip>=4.7.0,<4.8.0

# Cryptography and Security
cryptography>=40.0.0,<41.0.0
python-jose>=3.3.0,<3.4.0

# Database and Caching
redis>=4.5.0,<4.6.0
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0,<2.10.0

# Monitoring and Logging
sentry-sdk>=1.22.0,<1.23.0
prometheus-client>=0.16.0,<0.17.0

# Web and Networking
requests>=2.30.0,<2.31.0
websockets>=11.0.0,<11.1.0

# Development and Testing
pytest>=7.3.0,<7.4.0
black>=23.3.0,<23.4.0
mypy>=1.3.0,<1.4.0

# Performance Optimization
cython>=0.29.0,<0.30.0

# Environment and Configuration
click>=8.1.0,<8.2.0

# Cloud Provider Dependencies
boto3>=1.26.0
google-cloud-compute>=2.0.0
google-cloud-aiplatform>=1.25.0
azure-mgmt-compute>=29.1.0
azure-ai-ml>=1.9.0
python-digitalocean>=2.19.0
pyyaml>=6.0.1
